pyproject.toml
[project]
name = "storytelling-ua"
version = "0.1.0"
description = "Advanced story classification MCP server"
dependencies = [
    "fastmcp>=0.2.0",
    "aiohttp>=3.9.0",
    "beautifulsoup4>=4.12.0",
    "openpyxl>=3.1.0b1",
    "openai>=1.0.0",
    "requests>=2.31.0",
    "lxml>=4.9.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
include = [
    "main.py",
    "README.md",
]

[tool.uv]
dev-dependencies = []